refactor: Initialization, imports and README (#8)
- Removed init method for classes for KUMA components as initialization is being done in `KumaRestAPIModule`
- Changed imports from relative to absolute because it is considered best-practice
- Removed unused imports (from `typing` module)
- Translated annotations to ENG
- Changed `logging.py` to `_logging.py` because `logging` is reserved for module name
- Added datatypes annotations
- Added TODO
- Updated `README.md`
- Added specific versions for modules in `requirements.dev.txt`
