+++ README.md
@@ -30,7 +30,7 @@ import kuma
 
 # Инициализация клиента с публичным API
 client = kuma.RestClient(
-    base_url="https://kuma.example.com",
+    url="https://kuma.example.com",
     token="YOUR_BEARER_TOKEN",
     verify='core.cert'  # Путь к SSL-сертификату (рекомендуется для продакшена)
 )
@@ -145,4 +145,4 @@ if status != 200:
 Официальная документация по публичному API доступна по адресу:
 https://support.kaspersky.com/help/KUMA/3.4/en-US/217973.htm
 
-**Примечание:** Приватные API-методы не документированы и могут изменяться вендором без предупреждения между версиями системы.
+**Примечание:** Приватные API-методы не задокументированы и могут изменяться вендором без предупреждения между версиями системы.
