+++ kuma/__init__.py
@@ -1,4 +1,4 @@
-from .private import KumaPrivateAPI as PrivateClient
-from .rest import KumaRestAPI as RestClient
+from kuma.private import KumaPrivateAPI as PrivateClient
+from kuma.rest import KumaRestAPI as RestClient
 
 __all__ = ["PrivateClient", "RestClient"]
