+++ kuma/private.py
@@ -6,8 +6,8 @@ from typing import Any, Tuple
 
 import requests
 
-from .constants import SHARED_TENANT
-from .logging import configure_logging
+from kuma.constants import SHARED_TENANT
+from kuma._logging import configure_logging
 
 _logger = configure_logging()
 _api_version = "v2.1"
