+++ kuma/rest/assets.py
@@ -1,16 +1,10 @@
-from typing import Any, Dict, List, Optional, Tuple, Union
+from typing import Dict, List, Tuple
 
-from ._base import KumaRestAPIModule
+from kuma.rest._base import KumaRestAPIModule
 
 
 class KumaRestAPIAssets(KumaRestAPIModule):
-    """
-    Методы для работы с активами
-    """
-
-    def __init__(self, base):
-        super().__init__(base)
-
+    """Methods for Assets."""
     def search(self, **kwargs) -> Tuple[int, List | str]:
         """Searchin assets by provided filter
         Args:
