+++ kuma/rest/settings.py
@@ -1,16 +1,8 @@
-from typing import Dict, List, Optional, Tuple, Union
-
-from ._base import KumaRestAPIModule
+from kuma.rest._base import KumaRestAPIModule
 
 
 class KumaRestAPISettings(KumaRestAPIModule):
-    """
-    Методы для работы с настройками Ядра
-    """
-
-    def __init__(self, base):
-        super().__init__(base)
-
+    """Methods for Settings."""
     def view(self, id: str) -> tuple[int, dict | str]:
         """
         List of custom fields added by the KUMA user in the application web interface.
