+++ requirements.dev.txt
@@ -2,6 +2,6 @@ black==25.1.0
 isort==6.0.1
 pre-commit==4.2.0
 pytest==8.3.5
-setuptools
-wheel
-twine
+setuptools==80.9.0
+wheel==0.45.1
+twine==6.1.0
